Revert "Merge pull request #705 from Anto426/feature/large-artwork-library-sections"
This reverts commit a8793a402a810f33c17b9f4b5a00614579f6f1f2, reversing
changes made to 33c82f92fe1b55943a940035b5530166c05acebc.

+++ app/src/main/kotlin/com/music/vivi/utils/ComposeToImage.kt
@@ -115,7 +115,7 @@ object ComposeToImage {
                             
                             // Dark overlay for readability
                             val overlayPaint = Paint().apply {
-                                color = 0x4D000000 // 30% black overlay
+                                color = 0x4D000000.toInt() // 30% black overlay
                             }
                             canvas.drawRect(blurRect, overlayPaint)
                         }
